pyproject.toml
[project]
name = "nonebot-plugin-picstatus-template-zhenxun"
dynamic = ["version"]
description = "Template plugin project"
authors = [{ name = "LgCookie", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.4.3",
    "cookit[jinja,playwright,pydantic]>=0.13.1",
    "nonebot-plugin-picstatus>=2.1.3.post1",
    "nonebot-plugin-htmlrender>=0.6.6",
]
requires-python = ">=3.10,<4.0"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-picstatus-template-zhenxun"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "nonebot_plugin_picstatus_template_zhenxun/__init__.py"
